# pass zone into queue_captures