pulled out just the zone info per @Fryguy 

It is expensive to determine the zone, especially for
Storages that need to look up a number of objects.
Since it is known, just pass it into enqueue methods

TL;DR:

Running `Metric::Capture.perf_capture_timer` is 15% faster if zones are only looked up once.

https://bugzilla.redhat.com/show_bug.cgi?id=1227008

---

|VmOrTemplate|Storage|ExtManagementSystem|MiqRegion|Zone|
|        ---:|   ---:|               ---:|     ---:|---:|
|        100 |    13 |                 1 |       1 |  1 |

**before:** (master)

|     ms |  sql | sql ms | comments
|    ---:|  ---:|   ---:| ---
| 3344.1 | 1263 | 270.5 | a100-master-1
|  818.6 |  449 |  97.1 | a100-master-2
|  847.6 |  449 |  90.8 | a100-master-3
|  968.1 |  449 | 115.6 | a100-master-4

**after:** (this branch)

|      ms | sql |  sql ms | comments
|     ---:| ---:|    ---:| ---
|  3195.0 |1177 |  261.6 | a100-zone-1
|   742.8 | 363 |   89.8 | a100-zone-2
|   799.5 | 363 |   97.9 | a100-zone-3
|   708.0 | 363 |   84.7 | a100-zone-4

**averages for runs 2-4** (first run loads schema data and tends to be bloated)

metric|before|after|improvement
---|---|---|---
ms|877|749|15%
sql|449|363|19%
sql ms|101|90|10%