@jrafanie 2016-02-15
Nice @kbrock, this makes sense to me.

Note, I have no idea what the tables above mean... Is there a before/after there?  Is there a high level % queries saved (a out of b) for 100 vms, 13 storages, 1 ems, 1 server, 1 zone?

Maybe % time saved would be helpful too.