@kbrock 2016-02-16
I am concerned that I may be passing in zone for something that shouldn't have a zone set at all.

But it does look like all the places I passed it in, there was a default lookup for it anyway